+++ app/src/main/java/com/alchemtech/playlistmaker/creators/InternetCheckCreator.kt
@@ -1,13 +1,10 @@
-package com.alchemtech.playlistmaker.domain.creators
+package com.alchemtech.playlistmaker.creators
 
 import android.content.Context
 import com.alchemtech.playlistmaker.data.network.CheckInternetConnectionImpl
 
-class InternetCheckCreator(context: Context) {
-    val isChecked = checked(context)
-
-
-    private fun checked(context: Context): Boolean {
+object InternetCheckCreator {
+    fun provideInternetCheck(context: Context): Boolean {
         return CheckInternetConnectionImpl(context).hasInternetConnection()
     }
 }
