+++ app/src/main/java/com/alchemtech/playlistmaker/creators/ListTrackDbReadWriteCreator.kt
@@ -1,0 +1,13 @@
+package com.alchemtech.playlistmaker.creators
+
+import android.content.Context
+import com.alchemtech.playlistmaker.domain.db.ListTrackDbInteractor
+
+object ListTrackDbReadWriteCreator  {
+
+    fun provideListTrackDb(context: Context): ListTrackDbInteractor {
+        return ListTrackDbInteractor(context = context)
+
+    }
+
+}
