+++ app/src/main/java/com/alchemtech/playlistmaker/creators/PlayerDataFillingCreator.kt
@@ -1,0 +1,12 @@
+package com.alchemtech.playlistmaker.creators
+
+import android.content.Context
+import com.alchemtech.playlistmaker.databinding.ActivityPlayerBinding
+import com.alchemtech.playlistmaker.domain.entity.Track
+import com.alchemtech.playlistmaker.presentation.ui.player.PlayerFilling
+
+object PlayerDataFillingCreator {
+    fun provide( context: Context,binding: ActivityPlayerBinding,track: Track) : PlayerFilling {
+      return PlayerFilling(track, binding, context)
+    }
+}
