+++ app/src/main/java/com/alchemtech/playlistmaker/data/dto/repository/SharedPrefImpl.kt
@@ -1,14 +1,15 @@
-package com.alchemtech.playlistmaker.data.dto.sharePreferences
+package com.alchemtech.playlistmaker.data.dto.repository
 
 import android.content.Context
 import android.content.Context.MODE_PRIVATE
+import com.alchemtech.playlistmaker.domain.api.SharedPref
 import com.alchemtech.playlistmaker.data.dto.trackDto.TrackDto
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import java.io.Serializable
 
 interface SharedPrefImpl : SharedPref {
-    override fun getSavedPref(name: String, key: String , context: Context ): List<TrackDto>? {
+     override fun getSavedPref(name: String, key: String, context: Context ): List<TrackDto>? {
 
         val shared = context.getSharedPreferences(/* name = */ name, /* mode = */ MODE_PRIVATE)
         val json =
