+++ app/src/main/java/com/alchemtech/playlistmaker/domain/api/SharedPref.kt
@@ -1,11 +1,11 @@
-package com.alchemtech.playlistmaker.data.dto.sharePreferences
+package com.alchemtech.playlistmaker.domain.api
 
 import android.content.Context
 import com.alchemtech.playlistmaker.data.dto.trackDto.TrackDto
 import java.io.Serializable
 
 interface SharedPref {
-   fun getSavedPref(name:String, key : String, context: Context): List<TrackDto>?
+    fun getSavedPref(name:String, key : String, context: Context): List<TrackDto>?
     fun setSavedPref(name:String, key : String, objects: Serializable, context: Context)
 
 }
