+++ app/src/main/java/com/alchemtech/playlistmaker/domain/player/PlayerInteractor.kt
@@ -1,0 +1,14 @@
+package com.alchemtech.playlistmaker.domain.player
+
+/* объект заготовка */
+interface PlayerInteractor {
+    var playerState : Int
+
+    fun pausePlayer()
+
+    fun startPlayer()
+    fun playbackControl()
+
+    fun preparePlayer()
+    fun release()
+}
