+++ app/src/main/java/com/alchemtech/playlistmaker/domain/player/PlayerInteractorImplUseCase.kt
@@ -1,0 +1,112 @@
+package com.alchemtech.playlistmaker.domain.player
+
+import android.content.Context
+import android.media.MediaPlayer
+import android.os.Handler
+import android.os.Looper
+import com.alchemtech.playlistmaker.R
+import com.alchemtech.playlistmaker.databinding.ActivityPlayerBinding
+import com.alchemtech.playlistmaker.domain.entity.Track
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class PlayerInteractorImplUseCase(
+    val binding:ActivityPlayerBinding,
+    val context: Context,
+    val track: Track,
+) :
+    PlayerInteractor {
+    override var playerState: Int = STATE_DEFAULT
+    private var mediaPlayer = MediaPlayer()
+
+    private val currentPositionTask = createUpdateCurrentPositionTask()
+    var mainThreadHandler = Handler(Looper.getMainLooper())
+
+    init {
+        preparePlayer()
+    }
+
+    override fun pausePlayer() {
+        mediaPlayer.pause()
+        playerState = STATE_PAUSED
+        binding.playBut.setImageResource(R.drawable.play_but)
+        killCurrentPositionTask()
+        startGetCurrentPositionTask()
+    }
+
+    override fun startPlayer() {
+        mediaPlayer.start()
+        playerState = STATE_PLAYING
+        binding.playBut.setImageResource(R.drawable.pause_but)
+        startGetCurrentPositionTask()
+    }
+
+   override fun release() {
+        mediaPlayer.release()
+        killCurrentPositionTask()
+    }
+
+    override fun playbackControl() {
+        when (playerState) {
+            STATE_PLAYING -> {
+                pausePlayer()
+            }
+
+            STATE_PREPARED, STATE_PAUSED -> {
+                startPlayer()
+            }
+        }
+    }
+
+    override fun preparePlayer() {
+        mediaPlayer.setDataSource(track.previewUrl)
+        mediaPlayer.prepareAsync()
+        mediaPlayer.setOnPreparedListener {
+            playerState = STATE_PREPARED
+            binding.playBut.isEnabled = true
+        }
+        mediaPlayer.setOnCompletionListener {
+            killCurrentPositionTask()
+            binding.playTime.text = "00:00"
+            binding.playBut.setImageResource(R.drawable.play_but)
+            playerState = STATE_PREPARED
+        }
+    }
+
+    private fun createUpdateCurrentPositionTask(): Runnable {
+        return object : Runnable {
+            override fun run() {
+
+                val currentPosition = SimpleDateFormat(
+                    "mm:ss",
+                    Locale.getDefault()
+                ).format(mediaPlayer.currentPosition)
+
+                binding.playTime.text = currentPosition
+
+                mainThreadHandler.postDelayed(this, DEBOUNCE_GET_CURRENT_POSITION)
+
+            }
+        }
+    }
+
+    private fun startGetCurrentPositionTask() {
+        mainThreadHandler.post(
+            currentPositionTask
+        )
+    }
+
+    private fun killCurrentPositionTask() {
+        mainThreadHandler.removeCallbacks(
+            currentPositionTask
+        )
+    }
+
+    companion object {
+        private const val STATE_DEFAULT = 0
+        private const val STATE_PREPARED = 1
+        private const val STATE_PLAYING = 2
+        private const val STATE_PAUSED = 3
+        private const val DEBOUNCE_GET_CURRENT_POSITION = 300L
+    }
+}
