+++ app/src/main/java/com/alchemtech/playlistmaker/presentation/ui/player/PlayerFilling.kt
@@ -5,7 +5,7 @@ import android.view.View
 import android.widget.ImageView
 import com.alchemtech.playlistmaker.R
 import com.alchemtech.playlistmaker.databinding.ActivityPlayerBinding
-import com.alchemtech.playlistmaker.domain.models.Track
+import com.alchemtech.playlistmaker.domain.entity.Track
 import com.alchemtech.playlistmaker.presentation.ui.UiCalculator
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
