+++ app/src/main/java/com/alchemtech/playlistmaker/presentation/ui/tracks/TrackSearchAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.alchemtech.playlistmaker.R
-import com.alchemtech.playlistmaker.domain.models.Track
+import com.alchemtech.playlistmaker.domain.entity.Track
 
 class TrackSearchAdapter(private val trackListOf: List<Track>) :
     RecyclerView.Adapter<TrackCardViewHolder>() {
